refactor: removed duplicated code
